Cleanup

+++ pyathena/pop_synth/parsec.py
@@ -1,12 +1,10 @@
 import os
-import bs4
-import requests
 import zipfile
 from pathlib import Path
 import os.path as osp
 import numpy as np
 import pandas as pd
-from tqdm import tqdm
+from .downloader import downloader
 
 class PopSynthParsec(object):
 
@@ -16,15 +14,14 @@ class PopSynthParsec(object):
         'tracks': 'https://stev.oapd.inaf.it/PARSEC/Database/PARSECv2.0_VMS/all_tracks.zip'
     }
 
-    def __init__(self, rootdir=None, model_def='Z0.014', force_download=False):
+    def __init__(self, rootdir=None, model='Z0.014', force_download=False):
         if rootdir is None:
-            rootdir = Path(__file__).parent.absolute() / '../../data/pop_synth'
+            rootdir = Path(__file__).parent.absolute() / '../../data/pop_synth/parsec'
             if not rootdir.is_dir():
                 rootdir.mkdir(parents=True, exist_ok=True)
 
         self.rootdir = Path(rootdir)
         self.dirs = dict()
-        self.force_download = force_download
         self.files_zip = {
             'photons': self.rootdir / 'all_photons.zip',
             'ejecta': self.rootdir / 'all_ejecta.zip',
@@ -32,15 +29,20 @@ class PopSynthParsec(object):
         }
 
         # Download files and unzip
-        for k in self.files_zip.keys():
-            self.download_file(k, force_download=self.force_download)
-            self.dirs[k] = osp.splitext(self.files_zip[k])[0]
-            if not osp.exists(self.dirs[k]) or force_download:
+        for k, v in self.files_zip.items():
+            message = f'Downloading {k} tables from PAdova TRieste Stellar'\
+                ' Evolutionary Code.\n'\
+                'https://stev.oapd.inaf.it/PARSEC/tracks_v2_VMS.html'
+            downloader(v, PopSynthParsec.urls[k], message,
+                       force_download=force_download)
+            self.dirs[k] = self.files_zip[k].with_suffix('')
+            if not self.dirs[k].exists() or force_download:
+                print('Unzip files')
                 self.unzip_one(self.files_zip[k], self.dirs[k])
                 self.unzip_all(self.dirs[k])
 
         self._find_models_and_dirs()
-        self.set_model(model_def)
+        self.set_model(model)
 
     def set_model(self, model):
         if model not in self.models:
@@ -50,46 +52,44 @@ class PopSynthParsec(object):
         self.df['photons'] = self.dfa['photons'][model]
         self.df['tracks'] = self.dfa['tracks'][model]
 
-    def download_file(self, kind, force_download=False):
-        fname = self.files_zip[kind]
-        url = PopSynthParsec.urls[kind]
-        if not osp.exists(fname) or force_download:
-            print('Downloading {0:s} tables from PAdova TRieste Stellar Evolutionary Code.'.\
-                  format(kind))
-            print('https://stev.oapd.inaf.it/PARSEC/tracks_v2_VMS.html')
+    # def download_file(self, kind, force_download=False):
+    #     fname = self.files_zip[kind]
+    #     url = PopSynthParsec.urls[kind]
+    #     if not osp.exists(fname) or force_download:
+    #         print('Downloading {0:s} tables from PAdova TRieste Stellar Evolutionary Code.'.\
+    #               format(kind))
+    #         print('https://stev.oapd.inaf.it/PARSEC/tracks_v2_VMS.html')
 
-            response = requests.get(url, stream=True)
-            if not response.ok:
-                raise Exception('Failed to download file. Check URL.')
+    #         response = requests.get(url, stream=True)
+    #         if not response.ok:
+    #             raise Exception('Failed to download file. Check URL.')
 
-            total_size = int(response.headers.get('content-length', 0))  # Get file size
-            block_size = 1024
+    #         total_size = int(response.headers.get('content-length', 0))  # Get file size
+    #         block_size = 1024
 
-            with open(fname, 'wb') as f, tqdm(
-                desc='Downloading',
-                total=total_size,
-                unit='B',
-                unit_scale=True,
-                unit_divisor=1024  # Convert to KB/MB
-            ) as bar:
-                for chunk in response.iter_content(chunk_size=block_size):
-                    f.write(chunk)
-                    bar.update(len(chunk))  # Update progress bar
+    #         with open(fname, 'wb') as f, tqdm(
+    #             desc='Downloading',
+    #             total=total_size,
+    #             unit='B',
+    #             unit_scale=True,
+    #             unit_divisor=1024  # Convert to KB/MB
+    #         ) as bar:
+    #             for chunk in response.iter_content(chunk_size=block_size):
+    #                 f.write(chunk)
+    #                 bar.update(len(chunk))  # Update progress bar
 
     @staticmethod
-    def unzip_one(fname, extractdir=None, delete_zip=False):
+    def unzip_one(fname, extractdir, delete_zip=False):
         """
         Unzips a single ZIP file.
         """
-        if extractdir is None:
-            # Extract to a folder with the ZIP file's name
-            extractdir = osp.splitext(fname)[0]
-
-        os.makedirs(extractdir, exist_ok=True)  # Ensure output directory exists
+        # Ensure output directory exists
+        os.makedirs(extractdir, exist_ok=True)
+        if not zipfile.is_zipfile(fname):
+            raise IOError(f'{fname} is not a valid zip file. '
+                          'Try again with `force_download=True`')
         with zipfile.ZipFile(fname, 'r') as zip_ref:
             zip_ref.extractall(extractdir)
-            # if verbose:
-            #print(f'Extracted: {fname} -> {extractdir}')
 
         if delete_zip:
             os.remove(fname)
